# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bipupu",
    "description": "BIPUPU API 服务",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/profile/avatar": {
      "post": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Upload Avatar",
        "description": "上传并更新用户头像（存储到数据库）- 优化版本",
        "operationId": "upload_avatar_api_profile_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_profile_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/me": {
      "get": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Get Current User Info",
        "description": "获取当前用户信息",
        "operationId": "get_current_user_info_api_profile_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/": {
      "get": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Get User Profile",
        "description": "获取用户详细资料",
        "operationId": "get_user_profile_api_profile__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Update User Profile",
        "description": "更新用户详细资料",
        "operationId": "update_user_profile_api_profile__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/password": {
      "put": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Update Password",
        "description": "更新用户密码",
        "operationId": "update_password_api_profile_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPasswordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/timezone": {
      "put": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Update Timezone",
        "description": "更新用户时区设置",
        "operationId": "update_timezone_api_profile_timezone_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimezoneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/push-settings": {
      "get": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Get Push Settings",
        "description": "获取用户推送设置信息",
        "operationId": "get_push_settings_api_profile_push_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Push Settings Api Profile Push Settings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/avatar/{bipupu_id}": {
      "get": {
        "tags": [
          "用户资料",
          "用户资料"
        ],
        "summary": "Get User Avatar",
        "description": "获取用户头像（通过用户业务 ID bipupu_id 提供）- 优化版本",
        "operationId": "get_user_avatar_api_profile_avatar__bipupu_id__get",
        "parameters": [
          {
            "name": "bipupu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bipupu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_avatar_api_profile_avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_profile_avatar_post"
      },
...
      "CosmicProfile": {
        "properties": {
          "birthday": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday",
            "description": "公历生日，格式 YYYY-MM-DD"
          },
          "zodiac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zodiac",
            "description": "西方星座，例如 '白羊座'"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "年龄（整数岁）"
          },
          "bazi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bazi",
            "description": "生辰八字字符串（可由用户提供或第三方服务计算）"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "description": "性别，例如 'male','female','other'，可选"
          },
          "mbti": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbti",
            "description": "MBTI 类型，例如 'INTJ'"
          },
          "birth_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Time",
            "description": "出生时间，例如 '08:30'"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace",
            "description": "出生地，城市/经纬度等"
          }
        },
        "type": "object",
        "title": "CosmicProfile",
        "description": "结构化的宇宙信息（可在未来扩展）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "StatusResponse",
        "description": "通用状态响应"
      },
...
      "TimezoneUpdate": {
        "properties": {
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "description": "时区标识符，例如 'Asia/Shanghai', 'America/New_York', 'Europe/London'",
            "examples": [
              "Asia/Shanghai",
              "America/New_York",
              "Europe/London",
              "UTC"
            ]
          }
        },
        "type": "object",
        "required": [
          "timezone"
        ],
        "title": "TimezoneUpdate",
        "description": "时区更新"
      },
...
      "UserPasswordUpdate": {
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "UserPasswordUpdate",
        "description": "用户密码更新"
      },
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "bipupu_id": {
            "type": "string",
            "title": "Bipupu Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "cosmic_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CosmicProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "default": false
          },
          "last_active": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Active",
            "description": "最后活跃时间"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "username",
          "bipupu_id",
          "id",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "用户响应模式"
      },
...
      "UserUpdate": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "cosmic_profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cosmic Profile"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "更新用户模式"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}